Rename move infeed compose's package.
