+++ feature/home/compose/src/main/java/com/tfandkusu/androidview/compose/home/listitem/InfeedAdAndroidView.kt
@@ -1,4 +1,4 @@
-package com.tfandkusu.androidview.compose.home
+package com.tfandkusu.androidview.compose.home.listitem
 
 import android.view.View
 import androidx.compose.foundation.layout.Column
